requirements.txt
pyYaml >= 6.0